Rename errors to error
